ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
numpy==1.26.4
tensorflow-cpu==2.13.0
pillow==10.3.0
python-multipart==0.0.9